Move `parse_webhook` to `utils` module
